thermal-mgmt: Move power allocation to a separate file
This patch moves power allocation functions to a separate file.
It will allow to encapsulate this functionality.
