+++ module/thermal_mgmt/CMakeLists.txt
@@ -11,7 +11,8 @@ target_include_directories(${SCP_MODULE_TARGET}
                            PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/include")
 
 target_sources(${SCP_MODULE_TARGET}
-               PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/src/mod_thermal_mgmt.c")
+               PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/src/mod_thermal_mgmt.c"
+               PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/src/power_allocation.c")
 
 target_link_libraries(${SCP_MODULE_TARGET} PRIVATE module-dvfs
                                                    module-scmi-perf
